requirements.txt
discord.py
python-dotenv~=1.1.1